% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcs.tree.R
\name{pcs.tree}
\alias{pcs.tree}
\title{Solve the Prize-Collecting Steiner Tree problem}
\usage{
pcs.tree(graph, terminals, lambda, root, depth, eps, max_iter, terminal_infty=10000)
}
\arguments{
\item{graph}{an igraph graph}

\item{terminals}{a numeric or character vector.}

\item{lambda}{a numeric parameter.}

\item{root}{a numeric or character scalar.}

\item{depth}{a numeric scalar.}

\item{eps}{a numeric scalar.}

\item{max_iter}{a numeric scalar; specifies maximum number of iterations.}

\item{terminal_infty}{a numeric scalar.}
}
\value{
Returns a list with cost and edges of the final tree.
}
\description{
Solve the Prize-Collecting Steiner Tree problem.
}
\examples{
g <- graph('Bull')
E(g)$costs  <- c(3, 3, 3, 3, 3)
V(g)$prizes <- c(10, 2, 2, 2, 2)
treeData <- pcs.tree(graph=g, terminals=c(4,5), lambda=1, root=3, depth=5, eps=1e-3, max_iter=10)

}
\references{
1. M. Bayati, C. Borgs, A. Braunstein, J. Chayes, A. Ramezanpour, and R. Zecchina,
               "Statistical Mechanics of Steiner Trees". PRL, 2008.

            2. M. Bayati, A. Braunstein, and R. Zecchina, "A rigorous analysis of the cavity
               equations for the minimum spanning tree". Journal of Mathematical Physics, 2008.

            3. I. Biazzo, A. Braunstein and R. Zecchina, "Performance of a cavity-method-based
               algorithm for the prize-collecting Steiner tree problem on graphs". PRL, 2012.
}
